Reduce memory usage

+++ Assets/ScriptableObjects/Item.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class Item : ScriptableObject
 {
     public string itemName;
-    public int itemID;
+    public byte itemID;
     public Sprite itemImage;
     public bool stackable;
 }
